+++ packages/blade/src/components/Card/LinkOverlay.native.tsx
@@ -1,0 +1,6 @@
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const LinkOverlay = (_props: any): React.ReactElement => {
+  return <></>;
+};
+
+export { LinkOverlay };