+++ packages/blade/src/components/Typography/Display/index.ts
@@ -1,0 +1,2 @@
+export { Display } from './Display';
+export type { DisplayProps } from './Display';